fix utils scoping
