+++ torch/_dynamo/source.py
@@ -17,7 +17,6 @@ guard generation, and in code reconstruction by providing methods to rebuild
 the code needed to recreate values.
 """
 
-import contextvars
 import dataclasses
 import enum
 import functools
@@ -43,22 +42,6 @@ from .bytecode_transformation import (
 
 if TYPE_CHECKING:
     from .codegen import PyCodegen
-
-
-_CONTEXTVAR_EXPLICIT_STATE_SENTINEL = object()
-
-
-def _contextvar_has_explicit_binding(cv: contextvars.ContextVar[Any]) -> bool:
-    return (
-        contextvars.copy_context().get(cv, _CONTEXTVAR_EXPLICIT_STATE_SENTINEL)
-        is not _CONTEXTVAR_EXPLICIT_STATE_SENTINEL
-    )
-
-
-def _contextvar_get_explicit_value_or_missing(
-    cv: contextvars.ContextVar[Any],
-) -> Any:
-    return contextvars.copy_context().get(cv, contextvars.Token.MISSING)
 
 
 # It shouldn't be supported to construct an NNModuleVariable inside an FSDP module,
@@ -1267,7 +1250,7 @@ class ContextVarExplicitValueSource(ChainedSource):
     def reconstruct(self, codegen: "PyCodegen") -> None:
         codegen.add_push_null(
             lambda: codegen.load_import_from(
-                "torch._dynamo.source", "_contextvar_get_explicit_value_or_missing"
+                "torch._dynamo.utils", "_contextvar_get_explicit_value_or_missing"
             )
         )
         codegen(self.base)
@@ -1275,7 +1258,7 @@ class ContextVarExplicitValueSource(ChainedSource):
 
     @property
     def _name_template(self) -> str:
-        return "__import__('torch._dynamo.source', fromlist=['_contextvar_get_explicit_value_or_missing'])._contextvar_get_explicit_value_or_missing({0})"
+        return "__import__('torch._dynamo.utils', fromlist=['_contextvar_get_explicit_value_or_missing'])._contextvar_get_explicit_value_or_missing({0})"
 
 
 @dataclass_with_cached_hash(frozen=True)
@@ -1283,7 +1266,7 @@ class ContextVarExplicitStateSource(ChainedSource):
     def reconstruct(self, codegen: "PyCodegen") -> None:
         codegen.add_push_null(
             lambda: codegen.load_import_from(
-                "torch._dynamo.source", "_contextvar_has_explicit_binding"
+                "torch._dynamo.utils", "_contextvar_has_explicit_binding"
             )
         )
         codegen(self.base)
@@ -1291,7 +1274,7 @@ class ContextVarExplicitStateSource(ChainedSource):
 
     @property
     def _name_template(self) -> str:
-        return "__import__('torch._dynamo.source', fromlist=['_contextvar_has_explicit_binding'])._contextvar_has_explicit_binding({0})"
+        return "__import__('torch._dynamo.utils', fromlist=['_contextvar_has_explicit_binding'])._contextvar_has_explicit_binding({0})"
 
 
 # This is a synthetic source that is associated with the singleton
